implemented UrlDisplay with UrlFlock and UrlOverview
