+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ware",
-  "version": "0.1.11",
+  "version": "0.1.12",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
