+++ src/components/v2/urls.css
@@ -1,0 +1,97 @@
+.url-display {
+    width:25%;
+    /*margin-left:1rem;*/
+    overflow: auto;
+}
+.url-display a {
+    /*color:red;*/
+    /*display:block;*/
+    width:10em;
+    margin-bottom:.5rem;
+    overflow-wrap: break-word;
+}
+.url-flock {
+    margin-top: 1.2rem;
+}
+.url-list .url-row {
+    display: flex;
+    flex-direction: row;
+    border:1px solid black;
+    border-top:none;
+    position:relative;
+}
+.url-list .url-row:first-child {
+    border-top:1px solid black;
+}
+.url-list .url-row.url-is-unknown,
+.url-list .url-row.url-is-unknown a{
+    color:dimgrey !important;
+}
+.url-list .url-row.url-is-redirect,
+.url-list .url-row.url-is-redirect a{
+    /*color: #b700ff !important;*/
+}
+.url-list .url-row.url-is-notfound,
+.url-list .url-row.url-is-notfound a{
+    color: red !important;
+}
+.url-list .url-row.url-is-error,
+.url-list .url-row.url-is-error a{
+    color: red !important;
+}
+
+.url-list .url-name {
+    width: 80%;
+    min-width: 80%;
+    max-width: 80%;
+    padding: .5rem;
+    border-right: 1px solid black;
+}
+.url-list .url-status {
+    width: 10%;
+    padding:.25rem;
+}
+.url-list .url-header-row .url-status {
+    cursor: pointer;
+    width:100%;
+    text-align: center;
+    color:#0000EE;
+}
+.url-list a {
+    width:80%;
+    margin-bottom:.5rem;
+    overflow-wrap: break-word;
+}
+
+.url-chart-display {
+    width:317px;
+}
+
+
+.url-filters {
+    display:flex;
+    flex-d5irection: column;
+
+}
+.url-filters button {
+    /*width:100% !important;*/
+    font-size:1rem;
+    padding:.25rem .5rem;
+    text-align:left;
+    font-weight: bold;
+    /*background-color: #cee1e8;*/
+    min-width: 10rem;
+    height:2.5rem;
+}
+.url-filters button.btn-pressed {
+    /*background-color: #7fbfe3 !important;*/
+    background-color: #95c9e5 !important;
+}
+.url-filters .btn-description {
+    display:none;
+    font-style: italic;
+    font-size: smaller;
+    font-weight: normal;
+    padding-left:1em;
+}
+
