+++ src/index.css
@@ -162,3 +162,10 @@ table.single-detail tr td:not(:first-child) {
     margin-left:10px;
 }
 
+.section-box {
+    padding:.8rem;
+    border:1px solid black;
+}
+.section-box h3, .section-box h4 {
+    margin-top:0;
+}
